+++ core/capabilities/ccip/ccipaptos/crcwconfig.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 
-	aptosloop "github.com/smartcontractkit/chainlink-aptos/relayer/chainreader/loop"
+	aptosloop "github.com/smartcontractkit/chainlink-aptos/codec/loop"
 	"github.com/smartcontractkit/chainlink-common/pkg/types"
 	ccipcommon "github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/common"
 	aptosconfig "github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/configs/aptos"