+++ deployment/vault/changeset/ethbalmon_set_watchlist_test.go
@@ -173,7 +173,7 @@ func TestEthBalMonSetWatchListChangeset(t *testing.T) {
 
 	deployCfg := types.DeployEthBalMonInput{
 		Chains: map[uint64]types.DeployEthBalMonChainConfig{
-			selector: {SetKeeperRegistryAddress: testAddr1},
+			selector: {KeeperRegistryAddress: testAddr1},
 		},
 	}
 	deployTask := runtime.ChangesetTask(DeployEthBalMonChangeSet, deployCfg)